# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "小红书 MCP 服务的 REST API 接口文档\n\n提供完整的小红书内容发布、互动和数据分析功能\n\n## 功能模块\n- **登录认证**: 登录状态检查、二维码登录、Cookie 管理\n- **内容发布**: 图文笔记、视频笔记发布\n- **内容发现**: Feed 列表、搜索、笔记详情\n- **用户信息**: 用户主页、个人资料\n- **内容互动**: 评论、点赞、收藏、关注\n\n## MCP 协议\n本服务同时提供 MCP (Model Context Protocol) 接口供 AI 工具使用\n- MCP 端点: /mcp\n- REST 端点: /api/v1/*",
        "title": "小红书 MCP API",
        "contact": {
            "name": "API Support",
            "url": "https://github.com/xpzouying/xiaohongshu-mcp"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "2.0.0"
    },
    "host": "localhost:18060",
    "basePath": "/api/v1",
    "paths": {
        "/feeds/comment": {
            "post": {
                "description": "在小红书笔记下发表评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容互动"
                ],
                "summary": "发表评论",
                "parameters": [
                    {
                        "description": "评论请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PostCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "评论成功",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "评论失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feeds/comment/reply": {
            "post": {
                "description": "回复笔记下的指定评论",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容互动"
                ],
                "summary": "回复评论",
                "parameters": [
                    {
                        "description": "回复请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.ReplyCommentRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "回复成功",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "回复失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feeds/detail": {
            "post": {
                "description": "获取小红书笔记完整详情和评论列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发现"
                ],
                "summary": "获取笔记详情",
                "parameters": [
                    {
                        "description": "详情请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.FeedDetailRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "笔记详情",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feeds/list": {
            "get": {
                "description": "获取小红书首页推荐 Feed 流",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发现"
                ],
                "summary": "获取首页 Feeds 列表",
                "responses": {
                    "200": {
                        "description": "Feed 列表",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/feeds/search": {
            "get": {
                "description": "搜索小红书笔记，支持关键词和多维度筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发现"
                ],
                "summary": "搜索笔记",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词(GET)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "description": "搜索请求参数(POST)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.SearchFeedsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索结果",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "搜索失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "搜索小红书笔记，支持关键词和多维度筛选",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发现"
                ],
                "summary": "搜索笔记",
                "parameters": [
                    {
                        "type": "string",
                        "description": "搜索关键词(GET)",
                        "name": "keyword",
                        "in": "query"
                    },
                    {
                        "description": "搜索请求参数(POST)",
                        "name": "request",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/main.SearchFeedsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "搜索结果",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "搜索失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login/cookies": {
            "delete": {
                "description": "删除本地保存的 cookies 文件，重置登录状态",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录认证"
                ],
                "summary": "删除 Cookies",
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "删除失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login/qrcode": {
            "get": {
                "description": "生成小红书登录二维码，返回 Base64 编码的图片和超时时间",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录认证"
                ],
                "summary": "获取登录二维码",
                "responses": {
                    "200": {
                        "description": "二维码信息",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/login/status": {
            "get": {
                "description": "检查当前是否已登录小红书",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登录认证"
                ],
                "summary": "检查登录状态",
                "responses": {
                    "200": {
                        "description": "登录状态信息",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "服务器内部错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/publish": {
            "post": {
                "description": "发布小红书图文内容，支持标题、正文、图片、标签等",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发布"
                ],
                "summary": "发布图文笔记",
                "parameters": [
                    {
                        "description": "发布请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PublishRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "发布失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/publish_video": {
            "post": {
                "description": "发布小红书视频内容",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "内容发布"
                ],
                "summary": "发布视频笔记",
                "parameters": [
                    {
                        "description": "视频发布请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.PublishVideoRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "发布成功",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "发布失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/me": {
            "get": {
                "description": "获取当前登录用户的主页信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户信息"
                ],
                "summary": "获取我的资料",
                "responses": {
                    "200": {
                        "description": "用户信息",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/user/profile": {
            "post": {
                "description": "获取指定用户的主页信息和笔记列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户信息"
                ],
                "summary": "获取用户主页",
                "parameters": [
                    {
                        "description": "用户主页请求参数",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/main.UserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "用户主页信息",
                        "schema": {
                            "$ref": "#/definitions/main.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "请求参数错误",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "获取失败",
                        "schema": {
                            "$ref": "#/definitions/main.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.CommentLoadConfig": {
            "type": "object",
            "properties": {
                "click_more_replies": {
                    "description": "是否点击\"更多回复\"按钮",
                    "type": "boolean"
                },
                "max_comment_items": {
                    "description": "最大加载评论数（.parent-comment数量），0表示加载所有",
                    "type": "integer"
                },
                "max_replies_threshold": {
                    "description": "回复数量阈值，超过这个数量的\"更多\"按钮将被跳过（0表示不跳过任何）",
                    "type": "integer"
                },
                "scroll_speed": {
                    "description": "滚动速度等级: slow(慢速), normal(正常), fast(快速)",
                    "type": "string"
                }
            }
        },
        "main.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "string"
                }
            }
        },
        "main.FeedDetailRequest": {
            "type": "object",
            "required": [
                "feed_id",
                "xsec_token"
            ],
            "properties": {
                "comment_config": {
                    "$ref": "#/definitions/main.CommentLoadConfig"
                },
                "feed_id": {
                    "type": "string"
                },
                "load_all_comments": {
                    "type": "boolean"
                },
                "xsec_token": {
                    "type": "string"
                }
            }
        },
        "main.PostCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "feed_id",
                "xsec_token"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "xsec_token": {
                    "type": "string"
                }
            }
        },
        "main.PublishRequest": {
            "type": "object",
            "required": [
                "content",
                "images",
                "title"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "images": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "type": "string"
                    }
                },
                "schedule_at": {
                    "description": "定时发布时间，ISO8601格式，为空则立即发布",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "main.PublishVideoRequest": {
            "type": "object",
            "required": [
                "content",
                "title",
                "video"
            ],
            "properties": {
                "content": {
                    "type": "string"
                },
                "schedule_at": {
                    "description": "定时发布时间，ISO8601格式，为空则立即发布",
                    "type": "string"
                },
                "tags": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "title": {
                    "type": "string"
                },
                "video": {
                    "type": "string"
                }
            }
        },
        "main.ReplyCommentRequest": {
            "type": "object",
            "required": [
                "content",
                "feed_id",
                "xsec_token"
            ],
            "properties": {
                "comment_id": {
                    "type": "string"
                },
                "content": {
                    "type": "string"
                },
                "feed_id": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                },
                "xsec_token": {
                    "type": "string"
                }
            }
        },
        "main.SearchFeedsRequest": {
            "type": "object",
            "required": [
                "keyword"
            ],
            "properties": {
                "filters": {
                    "$ref": "#/definitions/xiaohongshu.FilterOption"
                },
                "keyword": {
                    "type": "string"
                }
            }
        },
        "main.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
        "main.UserProfileRequest": {
            "type": "object",
            "required": [
                "user_id",
                "xsec_token"
            ],
            "properties": {
                "user_id": {
                    "type": "string"
                },
                "xsec_token": {
                    "type": "string"
                }
            }
        },
        "xiaohongshu.FilterOption": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "note_type": {
                    "type": "string"
                },
                "publish_time": {
                    "type": "string"
                },
                "search_scope": {
                    "type": "string"
                },
                "sort_by": {
                    "type": "string"
                }
            }
        }
    },
...
}